Hadizia Beta V0
Contains ZIP file with Mod workspace file + test extracted mod itself. This is version zero, so don't expect it to be perfect. Report any bugs or opinions to the discord server. Fairies are not present in this current version. Please refer to the "README" file in the zip after download.